requirements.txt
torchvision==0.8.2
torchviz==0.0.2
tqdm==4.62.0
transformers==4.9.1
typing-extensions
urllib3==1.26.6
xxhash==2.0.2
